LayoutTests/fast/images/image-extraction/ios/select-word-in-image-overlay-inside-link-expected.txt

PASS selectionRects.length is 1
PASS successfullyParsed is true

TEST COMPLETE
